refactor: Remove inheritance from methods builder
- Remove inheritance from filtered methods builder and no methods
  builder
- Use the methods builder interface instead
