+++ src/Parser/Code/Builders/Members/MethodsBuilder.php
@@ -9,7 +9,6 @@ namespace PhUml\Parser\Code\Builders\Members;
 
 use PhpParser\Node\Stmt\ClassMethod;
 use PhUml\Code\Methods\Method;
-use PhUml\Code\Methods\MethodDocBlock;
 
 /**
  * It builds an array with `Method`s for a `ClassDefinition`, an `InterfaceDefinition` or a
@@ -20,58 +19,11 @@ use PhUml\Code\Methods\MethodDocBlock;
  * @see PrivateVisibilityFilter
  * @see ProtectedVisibilityFilter
  */
-class MethodsBuilder
+interface MethodsBuilder
 {
-    /** @var ParametersBuilder */
-    private $parametersBuilder;
-
-    /** @var VisibilityBuilder */
-    private $visibilityBuilder;
-
-    /** @var TypeBuilder */
-    private $typeBuilder;
-
-    /** @var VisibilityFilters */
-    private $visibilityFilters;
-
-    public function __construct(
-        ParametersBuilder $parametersBuilder,
-        TypeBuilder $typeBuilder,
-        VisibilityBuilder $visibilityBuilder,
-        VisibilityFilters $filters
-    ) {
-        $this->visibilityFilters = $filters;
-        $this->typeBuilder = $typeBuilder;
-        $this->parametersBuilder = $parametersBuilder;
-        $this->visibilityBuilder = $visibilityBuilder;
-    }
-
     /**
      * @param ClassMethod[] $methods
      * @return Method[]
      */
-    public function build(array $methods): array
-    {
-        return array_map(function (ClassMethod $method): Method {
-            return $this->buildMethod($method);
-        }, $this->visibilityFilters->apply($methods));
-    }
-
-    private function buildMethod(ClassMethod $method): Method
-    {
-        $name = $method->name->name;
-        $visibility = $this->visibilityBuilder->build($method);
-        $docBlock = $method->getDocComment() === null ? null : $method->getDocComment()->getText();
-        $methodDocBlock = MethodDocBlock::from($docBlock);
-        $returnType = $this->typeBuilder->fromMethodReturnType($method->returnType, $methodDocBlock);
-        $parameters = $this->parametersBuilder->build($method->params, $methodDocBlock);
-        switch (true) {
-            case $method->isAbstract():
-                return new Method($name, $visibility, $returnType, $parameters, true);
-            case $method->isStatic():
-                return new Method($name, $visibility, $returnType, $parameters, false, true);
-            default:
-                return new Method($name, $visibility, $returnType, $parameters);
-        }
-    }
+    public function build(array $methods): array;
 }
